adding ability to perform crud with categories

+++ src/services/category.services.js
@@ -22,26 +22,37 @@ const getAllCategories = async () => {
 
 
 //editing category
-const editCategory = async (body, id) => {
+const editCategory = async (id, body) => {
+  // checking Category exists
+  const { rows } = await db.query('SELECT * FROM "categories" WHERE id = $1', [id]);
+  if (!rows[0]) {
+    throw new Error('Category does not exist');
+  }
   const { name, group_id } = body;
 
-  const { rows } = await db.query(
-    'UPDATE categories SET name = $1, group_id = $2 WHERE id = $3 RETURNING *',
+  const { rows: editedCategory } = await db.query(
+    'UPDATE "categories" SET name = $1, group_id = $2 WHERE id = $3 RETURNING *',
     [name, group_id, id]
   );
 
-  return rows[0];
+  return editedCategory[0];
 }
 
 
 //delete category
 const deleteCategory = async (id) => {
-  const { rows } = await db.query(
+  // checking category exists
+  const { rows } = await db.query('SELECT * FROM "categories" WHERE id = $1', [id]);
+  if (!rows[0]) {
+    throw new Error(`Category does not exist`);
+  }
+
+  const { rows: deletedCategory } = await db.query(
     'DELETE FROM categories WHERE id = $1 RETURNING *',
     [id]
   );
 
-  return rows[0];
+  return deletedCategory[0];
 }
 
 module.exports = {
